requirements.txt
pinecone-client==2.2.4
openai==1.12.0
python-dotenv==1.0.0
streamlit==1.30.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
pypdf==3.17.1
tiktoken==0.5.2 